# packages/api-client/openapi.json
{"openapi": "3.1.0", "info": {"title": "Gulp API", "version": "0.1.0"}, "paths": {"/capture": {"post": {"tags": ["capture"], "summary": "Capture", "operationId": "capture_capture_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CaptureRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CaptureResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}": {"get": {"tags": ["capture"], "summary": "Get Snapshot", "operationId": "get_snapshot_snapshots__snapshot_id__get", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SnapshotOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/cards/generate": {"post": {"tags": ["cards"], "summary": "Generate Cards Route", "operationId": "generate_cards_route_snapshots__snapshot_id__cards_generate_post", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SnapshotOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/cards/import": {"post": {"tags": ["cards"], "summary": "Import Cards Route", "operationId": "import_cards_route_snapshots__snapshot_id__cards_import_post", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CardsPayload"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/CardOut"}, "title": "Response Import Cards Route Snapshots  Snapshot Id  Cards Import Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/cards": {"get": {"tags": ["cards"], "summary": "List Cards Route", "operationId": "list_cards_route_snapshots__snapshot_id__cards_get", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/CardOut"}, "title": "Response List Cards Route Snapshots  Snapshot Id  Cards Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/cards/{card_id}": {"patch": {"tags": ["cards"], "summary": "Update Card Route", "operationId": "update_card_route_snapshots__snapshot_id__cards__card_id__patch", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}, {"name": "card_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Card Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CardPatch"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CardOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["cards"], "summary": "Delete Card Route", "operationId": "delete_card_route_snapshots__snapshot_id__cards__card_id__delete", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}, {"name": "card_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Card Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/export": {"post": {"tags": ["export"], "summary": "Export Snapshot", "operationId": "export_snapshot_snapshots__snapshot_id__export_post", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SnapshotOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/job": {"get": {"tags": ["export"], "summary": "Download Job", "operationId": "download_job_snapshots__snapshot_id__job_get", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/import": {"post": {"tags": ["export"], "summary": "Import Snapshot", "operationId": "import_snapshot_snapshots__snapshot_id__import_post", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_import_snapshot_snapshots__snapshot_id__import_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SnapshotOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/inbox": {"get": {"tags": ["inbox"], "summary": "Inbox", "operationId": "inbox_inbox_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InboxOut"}}}}}}}, "/snapshots/{snapshot_id}/pack": {"get": {"tags": ["pack"], "summary": "Get Pack", "operationId": "get_pack_snapshots__snapshot_id__pack_get", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PackOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/blocks/{block_id}": {"patch": {"tags": ["pack"], "summary": "Update Block Route", "operationId": "update_block_route_snapshots__snapshot_id__blocks__block_id__patch", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}, {"name": "block_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Block Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BlockUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"oneOf": [{"$ref": "#/components/schemas/ProseBlockOut"}, {"$ref": "#/components/schemas/FormulaBlockOut"}, {"$ref": "#/components/schemas/TableBlockOut"}, {"$ref": "#/components/schemas/FigureBlockOut"}, {"$ref": "#/components/schemas/ListBlockOut"}], "discriminator": {"propertyName": "type", "mapping": {"prose": "#/components/schemas/ProseBlockOut", "formula": "#/components/schemas/FormulaBlockOut", "table": "#/components/schemas/TableBlockOut", "figure": "#/components/schemas/FigureBlockOut", "list": "#/components/schemas/ListBlockOut"}}, "title": "Response Update Block Route Snapshots  Snapshot Id  Blocks  Block Id  Patch"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["pack"], "summary": "Delete Block Route", "operationId": "delete_block_route_snapshots__snapshot_id__blocks__block_id__delete", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}, {"name": "block_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Block Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/sections/{section_id}/blocks": {"post": {"tags": ["pack"], "summary": "Create Block Route", "operationId": "create_block_route_snapshots__snapshot_id__sections__section_id__blocks_post", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}, {"name": "section_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Section Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BlockCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"oneOf": [{"$ref": "#/components/schemas/ProseBlockOut"}, {"$ref": "#/components/schemas/FormulaBlockOut"}, {"$ref": "#/components/schemas/TableBlockOut"}, {"$ref": "#/components/schemas/FigureBlockOut"}, {"$ref": "#/components/schemas/ListBlockOut"}], "discriminator": {"propertyName": "type", "mapping": {"prose": "#/components/schemas/ProseBlockOut", "formula": "#/components/schemas/FormulaBlockOut", "table": "#/components/schemas/TableBlockOut", "figure": "#/components/schemas/FigureBlockOut", "list": "#/components/schemas/ListBlockOut"}}, "title": "Response Create Block Route Snapshots  Snapshot Id  Sections  Section Id  Blocks Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/blocks/{block_id}/messages": {"get": {"tags": ["pack"], "summary": "List Block Messages Route", "operationId": "list_block_messages_route_snapshots__snapshot_id__blocks__block_id__messages_get", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}, {"name": "block_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Block Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/MessageOut"}, "title": "Response List Block Messages Route Snapshots  Snapshot Id  Blocks  Block Id  Messages Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["pack"], "summary": "Post Block Message Route", "operationId": "post_block_message_route_snapshots__snapshot_id__blocks__block_id__messages_post", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}, {"name": "block_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Block Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/snapshots/{snapshot_id}/process": {"post": {"tags": ["processing"], "summary": "Process Snapshot Endpoint", "operationId": "process_snapshot_endpoint_snapshots__snapshot_id__process_post", "parameters": [{"name": "snapshot_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Snapshot Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SnapshotOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Health Health Get"}}}}}}}}, "components": {"schemas": {"BlockCreate": {"properties": {"content": {"oneOf": [{"$ref": "#/components/schemas/ProseWrite"}, {"$ref": "#/components/schemas/FormulaWrite"}, {"$ref": "#/components/schemas/TableWrite"}, {"$ref": "#/components/schemas/FigureWrite"}, {"$ref": "#/components/schemas/ListWrite"}], "title": "Content", "discriminator": {"propertyName": "type", "mapping": {"figure": "#/components/schemas/FigureWrite", "formula": "#/components/schemas/FormulaWrite", "list": "#/components/schemas/ListWrite", "prose": "#/components/schemas/ProseWrite", "table": "#/components/schemas/TableWrite"}}}, "position": {"type": "integer", "title": "Position"}}, "type": "object", "required": ["content", "position"], "title": "BlockCreate"}, "BlockUpdate": {"properties": {"content": {"anyOf": [{"oneOf": [{"$ref": "#/components/schemas/ProseWrite"}, {"$ref": "#/components/schemas/FormulaWrite"}, {"$ref": "#/components/schemas/TableWrite"}, {"$ref": "#/components/schemas/FigureWrite"}, {"$ref": "#/components/schemas/ListWrite"}], "discriminator": {"propertyName": "type", "mapping": {"figure": "#/components/schemas/FigureWrite", "formula": "#/components/schemas/FormulaWrite", "list": "#/components/schemas/ListWrite", "prose": "#/components/schemas/ProseWrite", "table": "#/components/schemas/TableWrite"}}}, {"type": "null"}], "title": "Content"}, "position": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Position"}}, "type": "object", "title": "BlockUpdate"}, "Body_import_snapshot_snapshots__snapshot_id__import_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_import_snapshot_snapshots__snapshot_id__import_post"}, "CaptureRequest": {"properties": {"url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Url"}, "text": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Text"}, "note": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Note"}, "title": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Title"}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "default": []}, "captured_via": {"$ref": "#/components/schemas/CapturedVia", "default": "in_app"}}, "type": "object", "title": "CaptureRequest"}, "CaptureResponse": {"properties": {"snapshot": {"$ref": "#/components/schemas/SnapshotOut"}, "duplicate": {"type": "boolean", "title": "Duplicate"}}, "type": "object", "required": ["snapshot", "duplicate"], "title": "CaptureResponse"}, "CapturedVia": {"type": "string", "enum": ["share_sheet", "wechat", "email", "in_app", "paste", "manual", "screenshot", "audio_memo"], "title": "CapturedVia"}, "CardDraft": {"properties": {"card_type": {"$ref": "#/components/schemas/CardType"}, "prompt": {"type": "string", "minLength": 1, "title": "Prompt"}, "answer": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Answer"}, "explanation": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Explanation"}, "options": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Options"}}, "type": "object", "required": ["card_type", "prompt"], "title": "CardDraft"}, "CardOrigin": {"type": "string", "enum": ["pack", "conversation", "user", "imported"], "title": "CardOrigin"}, "CardOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "card_type": {"$ref": "#/components/schemas/CardType"}, "prompt": {"type": "string", "title": "Prompt"}, "answer": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Answer"}, "explanation": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Explanation"}, "options": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Options"}, "origin": {"$ref": "#/components/schemas/CardOrigin"}, "status": {"$ref": "#/components/schemas/CardStatus"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "card_type", "prompt", "answer", "explanation", "options", "origin", "status", "created_at", "updated_at"], "title": "CardOut"}, "CardPatch": {"properties": {"status": {"anyOf": [{"$ref": "#/components/schemas/CardStatus"}, {"type": "null"}]}, "prompt": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Prompt"}, "answer": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Answer"}, "explanation": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Explanation"}, "options": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Options"}}, "type": "object", "title": "CardPatch", "description": "Partial update; provided content fields are re-validated per card type."}, "CardStatus": {"type": "string", "enum": ["draft", "accepted", "rejected"], "title": "CardStatus"}, "CardType": {"type": "string", "enum": ["short_answer", "mcq", "cloze", "explain", "apply", "recall"], "title": "CardType"}, "CardsPayload": {"properties": {"cards": {"items": {"$ref": "#/components/schemas/CardDraft"}, "type": "array", "maxItems": 100, "minItems": 1, "title": "Cards"}}, "type": "object", "required": ["cards"], "title": "CardsPayload"}, "CardsStatus": {"type": "string", "enum": ["generating", "ready", "failed"], "title": "CardsStatus", "description": "Inline card-generation job state; null = never triggered. Imports don't touch it."}, "FigureBlockOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "type": {"type": "string", "const": "figure", "title": "Type", "default": "figure"}, "label": {"type": "string", "title": "Label"}, "explanation": {"type": "string", "title": "Explanation"}}, "type": "object", "required": ["id", "label", "explanation"], "title": "FigureBlockOut"}, "FigureWrite": {"properties": {"type": {"type": "string", "const": "figure", "title": "Type", "default": "figure"}, "label": {"type": "string", "title": "Label"}, "explanation": {"type": "string", "title": "Explanation"}}, "type": "object", "required": ["label", "explanation"], "title": "FigureWrite"}, "FormulaBlockOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "type": {"type": "string", "const": "formula", "title": "Type", "default": "formula"}, "latex": {"type": "string", "title": "Latex"}, "explanation": {"type": "string", "title": "Explanation"}}, "type": "object", "required": ["id", "latex", "explanation"], "title": "FormulaBlockOut"}, "FormulaWrite": {"properties": {"type": {"type": "string", "const": "formula", "title": "Type", "default": "formula"}, "latex": {"type": "string", "title": "Latex"}, "explanation": {"type": "string", "title": "Explanation"}}, "type": "object", "required": ["latex", "explanation"], "title": "FormulaWrite"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "InboxOut": {"properties": {"items": {"items": {"$ref": "#/components/schemas/SnapshotOut"}, "type": "array", "title": "Items"}, "count": {"type": "integer", "title": "Count"}}, "type": "object", "required": ["items", "count"], "title": "InboxOut"}, "ListBlockOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "type": {"type": "string", "const": "list", "title": "Type", "default": "list"}, "items": {"items": {"type": "string"}, "type": "array", "title": "Items"}, "ordered": {"type": "boolean", "title": "Ordered", "default": false}}, "type": "object", "required": ["id", "items"], "title": "ListBlockOut"}, "ListWrite": {"properties": {"type": {"type": "string", "const": "list", "title": "Type", "default": "list"}, "items": {"items": {"type": "string"}, "type": "array", "title": "Items"}, "ordered": {"type": "boolean", "title": "Ordered", "default": false}}, "type": "object", "required": ["items"], "title": "ListWrite"}, "MediaType": {"type": "string", "enum": ["article", "pdf", "video", "podcast", "note", "screenshot", "audio", "webpage"], "title": "MediaType"}, "MessageCreate": {"properties": {"content": {"type": "string", "title": "Content"}}, "type": "object", "required": ["content"], "title": "MessageCreate"}, "MessageOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "role": {"type": "string", "title": "Role"}, "content": {"type": "string", "title": "Content"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}}, "type": "object", "required": ["id", "role", "content", "created_at"], "title": "MessageOut"}, "PackOut": {"properties": {"snapshot_id": {"type": "string", "format": "uuid", "title": "Snapshot Id"}, "status": {"$ref": "#/components/schemas/PackStatus"}, "title": {"type": "string", "title": "Title"}, "core_contributions": {"items": {"type": "string"}, "type": "array", "title": "Core Contributions"}, "key_insight": {"type": "string", "title": "Key Insight"}, "sections": {"items": {"$ref": "#/components/schemas/PackSectionOut"}, "type": "array", "title": "Sections"}, "references": {"items": {"$ref": "#/components/schemas/PackReferenceOut"}, "type": "array", "title": "References"}}, "type": "object", "required": ["snapshot_id", "status", "title", "core_contributions", "key_insight", "sections", "references"], "title": "PackOut"}, "PackReferenceOut": {"properties": {"citation": {"type": "string", "title": "Citation"}, "why_interesting": {"type": "string", "title": "Why Interesting"}}, "type": "object", "required": ["citation", "why_interesting"], "title": "PackReferenceOut"}, "PackSectionOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "heading": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Heading"}, "blocks": {"items": {"oneOf": [{"$ref": "#/components/schemas/ProseBlockOut"}, {"$ref": "#/components/schemas/FormulaBlockOut"}, {"$ref": "#/components/schemas/TableBlockOut"}, {"$ref": "#/components/schemas/FigureBlockOut"}, {"$ref": "#/components/schemas/ListBlockOut"}], "discriminator": {"propertyName": "type", "mapping": {"figure": "#/components/schemas/FigureBlockOut", "formula": "#/components/schemas/FormulaBlockOut", "list": "#/components/schemas/ListBlockOut", "prose": "#/components/schemas/ProseBlockOut", "table": "#/components/schemas/TableBlockOut"}}}, "type": "array", "title": "Blocks"}}, "type": "object", "required": ["id", "heading", "blocks"], "title": "PackSectionOut"}, "PackStatus": {"type": "string", "enum": ["generating", "ready"], "title": "PackStatus"}, "ProseBlockOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "type": {"type": "string", "const": "prose", "title": "Type", "default": "prose"}, "content": {"type": "string", "title": "Content"}}, "type": "object", "required": ["id", "content"], "title": "ProseBlockOut"}, "ProseWrite": {"properties": {"type": {"type": "string", "const": "prose", "title": "Type", "default": "prose"}, "content": {"type": "string", "title": "Content"}}, "type": "object", "required": ["content"], "title": "ProseWrite"}, "SnapshotOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "kind": {"$ref": "#/components/schemas/SourceKind"}, "title": {"type": "string", "title": "Title"}, "note": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Note"}, "status": {"$ref": "#/components/schemas/SnapshotStatus"}, "media_type": {"anyOf": [{"$ref": "#/components/schemas/MediaType"}, {"type": "null"}]}, "origin_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Origin Url"}, "content_body": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Content Body"}, "captured_via": {"anyOf": [{"$ref": "#/components/schemas/CapturedVia"}, {"type": "null"}]}, "cards_status": {"anyOf": [{"$ref": "#/components/schemas/CardsStatus"}, {"type": "null"}]}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "kind", "title", "note", "status", "media_type", "origin_url", "content_body", "captured_via", "cards_status", "tags", "created_at", "updated_at"], "title": "SnapshotOut"}, "SnapshotStatus": {"type": "string", "enum": ["queued", "unprocessed", "processing", "ready", "exported", "awaiting_review", "in_library", "needs_attention"], "title": "SnapshotStatus"}, "SourceKind": {"type": "string", "enum": ["snapshot", "conversation", "subscription"], "title": "SourceKind"}, "TableBlockOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "type": {"type": "string", "const": "table", "title": "Type", "default": "table"}, "headers": {"items": {"type": "string"}, "type": "array", "title": "Headers"}, "rows": {"items": {"items": {"type": "string"}, "type": "array"}, "type": "array", "title": "Rows"}, "caption": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Caption"}}, "type": "object", "required": ["id", "headers", "rows"], "title": "TableBlockOut"}, "TableWrite": {"properties": {"type": {"type": "string", "const": "table", "title": "Type", "default": "table"}, "headers": {"items": {"type": "string"}, "type": "array", "title": "Headers"}, "rows": {"items": {"items": {"type": "string"}, "type": "array"}, "type": "array", "title": "Rows"}, "caption": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Caption"}}, "type": "object", "required": ["headers", "rows"], "title": "TableWrite"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}